[refactor](SearchForm.tsx) Moved interfaces and converter to separate file.
